@@ -1,6 +1,6 @@
+use crate::FRAME_LENGTH;
 use crate::double_buffer::DoubleBuffer;
 use crate::socket_stream::SocketStream;
-use crate::FRAME_LENGTH;
 
 pub type Frame = [u8; FRAME_LENGTH];
 pub static FRAME_BUFFER: DoubleBuffer = DoubleBuffer::new();
@@ -18,7 +18,7 @@ pub fn get_frame(is_recording: bool) -> Option<[u8; 39040]> {
     None
 }
 pub fn send_frame(fb: [u8; 39040], stream: &mut SocketStream) -> bool {
-    if let Err(_) = stream.write_all(&fb) {
+    if stream.write_all(&fb).is_err() {
         return false;
     }
     stream.flush().is_ok()
